To ease autotests debugging: 		1)decrease retry failed count to 0 		2)erase video and logs directory before new suite run. 	If autotests run on local machine.

+++ src/main/java/utils/BeforeAfter.java
@@ -8,6 +8,7 @@ import org.testng.annotations.BeforeSuite;
 import java.io.File;
 import java.io.IOException;
 
+import static data.Flags.isLocal;
 import static utils.Logs.createFolder;
 import static utils.Video.moveVideo;
 
@@ -23,10 +24,13 @@ public class BeforeAfter {
     }
 
     public static void deleteDirectories() throws IOException {
+
         File sourceDirectory = new File("video");
         FileUtils.deleteDirectory(sourceDirectory);
+        if(isLocal()){
         sourceDirectory = new File("videoAndLogs");
         FileUtils.deleteDirectory(sourceDirectory);
+        }
     }
 
     @BeforeSuite
